+++ nature/opts/init.lua
@@ -20,7 +20,10 @@ local function setupOpt(name, default)
 end
 
 local defaultOpts = {
-	autocd = false
+	autocd = false,
+	greeting = string.format([[Welcome to {magenta}Hilbish{reset}, {cyan}%s{reset}.
+The nice lil shell for {blue}Lua{reset} fanatics!
+]], hilbish.user)
 }
 
 for optsName, default in pairs(defaultOpts) do